Add files via upload
This project focuses on solving the Traveling Salesman Problem (TSP) using Simulated Annealing (SA) in C++. TSP involves finding the shortest tour that visits all cities once and returns to the starting point, with travel costs being symmetric and defined as Euclidean distances. Due to the complexity of TSP, SA is used as a metaheuristic optimization method. The program starts with an initial random tour (solution \(S\)) and iteratively generates a neighbor solution \(S'\) by swapping two cities' positions in the tour. If \(S'\) has a lower cost, it is accepted; otherwise, it is accepted probabilistically, with the probability decreasing over iterations. The algorithm runs for 10,000 iterations to approximate the optimal TSP tour with the least cost.